Move calculations to a separate file
