+++ src/check-condition.test.ts
@@ -1,64 +1,6 @@
 import { ethers } from 'ethers';
-import {
-  calculateUpdateInPercentage,
-  checkSignedDataFreshness,
-  checkOnchainDataFreshness,
-  checkUpdateCondition,
-  HUNDRED_PERCENT,
-} from './check-condition';
-import { State, updateState } from './state';
+import { checkSignedDataFreshness, checkOnchainDataFreshness, checkUpdateCondition } from './check-condition';
 import { getUnixTimestamp, validSignedData } from '../test/fixtures';
-
-describe('calculateUpdateInPercentage', () => {
-  beforeEach(() => {
-    updateState(() => ({ logOptions: {} } as unknown as State));
-  });
-
-  it('calculates zero change', () => {
-    const updateInPercentage = calculateUpdateInPercentage(ethers.BigNumber.from(10), ethers.BigNumber.from(10));
-    expect(updateInPercentage).toEqual(ethers.BigNumber.from(0 * HUNDRED_PERCENT));
-  });
-
-  it('calculates 100 percent change', () => {
-    const updateInPercentage = calculateUpdateInPercentage(ethers.BigNumber.from(10), ethers.BigNumber.from(20));
-    expect(updateInPercentage).toEqual(ethers.BigNumber.from(1 * HUNDRED_PERCENT));
-  });
-
-  it('calculates positive to negative change', () => {
-    const updateInPercentage = calculateUpdateInPercentage(ethers.BigNumber.from(10), ethers.BigNumber.from(-5));
-    expect(updateInPercentage).toEqual(ethers.BigNumber.from(1.5 * HUNDRED_PERCENT));
-  });
-
-  it('calculates negative to positive change', () => {
-    const updateInPercentage = calculateUpdateInPercentage(ethers.BigNumber.from(-5), ethers.BigNumber.from(5));
-    expect(updateInPercentage).toEqual(ethers.BigNumber.from(2 * HUNDRED_PERCENT));
-  });
-
-  it('calculates initial zero to positive change', () => {
-    const updateInPercentage = calculateUpdateInPercentage(ethers.BigNumber.from(0), ethers.BigNumber.from(5));
-    expect(updateInPercentage).toEqual(ethers.BigNumber.from(5 * HUNDRED_PERCENT));
-  });
-
-  it('calculates initial zero to negative change', () => {
-    const updateInPercentage = calculateUpdateInPercentage(ethers.BigNumber.from(0), ethers.BigNumber.from(-5));
-    expect(updateInPercentage).toEqual(ethers.BigNumber.from(5 * HUNDRED_PERCENT));
-  });
-
-  it('calculates initial positive to zero change', () => {
-    const updateInPercentage = calculateUpdateInPercentage(ethers.BigNumber.from(5), ethers.BigNumber.from(0));
-    expect(updateInPercentage).toEqual(ethers.BigNumber.from(1 * HUNDRED_PERCENT));
-  });
-
-  it('calculates initial negative to zero change', () => {
-    const updateInPercentage = calculateUpdateInPercentage(ethers.BigNumber.from(-5), ethers.BigNumber.from(0));
-    expect(updateInPercentage).toEqual(ethers.BigNumber.from(1 * HUNDRED_PERCENT));
-  });
-
-  it('calculates initial negative to negative change', () => {
-    const updateInPercentage = calculateUpdateInPercentage(ethers.BigNumber.from(-5), ethers.BigNumber.from(-1));
-    expect(updateInPercentage).toEqual(ethers.BigNumber.from(0.8 * HUNDRED_PERCENT));
-  });
-});
 
 describe('checkUpdateCondition', () => {
   const onChainValue = ethers.BigNumber.from(500);
